This is the code for identifing vortecies
I tested it using image number 502, which shows a clear vortex between the two cylinders. After adjusting the threshold and other parameters, I could detect the vortex.

The main goal of the code is to calculate the frequency from the PIV dataset. My initial idea was to select a specific area in the image, run the analysis across all the photos, and count how many times the vortex passes through this window. By averaging over time, we could estimate the frequency.

However, since the current implementation processes one image at a time, it seems to be taking too long. Could you provide your input on this approach? Or should we try my method "for science"?